Don't look up or set the uid/gid
Apparently having the name set allows doing the id look up on the
machine where the tar is unpacked, making it totally symbolic.  This
is really the behavior we want since the user can put in an arbitrary
name and if that user exists on the system where the package is
installed everything comes out right.

It's also more portable not to shell out to `id`